requirements.txt
pynacl
cryptography
fastecdsa
ecdsa
ed25519
libnacl